+++ src/acconeer/exptool/app/new/plugin_loader.py
@@ -7,6 +7,7 @@ from .app_model import PluginSpec
 
 
 def load_default_plugins() -> list[PluginSpec]:
+    from acconeer.exptool.a121.algo.bilateration._plugin import BILATERATION_PLUGIN
     from acconeer.exptool.a121.algo.distance._detector_plugin import DISTANCE_DETECTOR_PLUGIN
     from acconeer.exptool.a121.algo.phase_tracking._plugin import PHASE_TRACKING_PLUGIN
     from acconeer.exptool.a121.algo.presence._detector_plugin import PRESENCE_DETECTOR_PLUGIN
@@ -21,4 +22,5 @@ def load_default_plugins() -> list[PluginSpec]:
         PRESENCE_DETECTOR_PLUGIN,
         VIRTUAL_BUTTON_PLUGIN,
         VIBRATION_PLUGIN,
+        BILATERATION_PLUGIN,
     ]